Export h/0,2 programatically
The -export() attribute in pt_docsh.hrl made it impossible to generate
embedded documentation for a whole project with just a single directive
in rebar.config, as the generated helpers would not be exported.
Now, when the export attribute is generated by the transformation
this is no longer the case and {parse_transform, pt_docsh} in
rebar.config is sufficient to build docs for the whole project.