Use a core transform to embed helper code
`h/0,2` helpers are now embedded in a Core Transform pass.
This allows to write the code in a regular Erlang module - `docsh_rt`
(_RunTime_ support).
Thanks to the fact that all relevant functions in `docsh_rt` are inlined,
`docsh_rt` can be used as a set of Core Erlang AST templates to generate
the final `h/0,2` helpers for each instrumented module.

The code embedded into final modules uses template functions
from `docsh_rt`, but partially applies them to freeze module name,
which is an explicit parameter for `docsh_rt:h/1,3`,
but would be inconvenient to type in the shell each time.