wip: Add docsh module doc
This is a WIP because this doc would look the best if it presented
examples which actually differ for h/1,2,3 and s/1,2,3.
This is hard now, as the edoc reader is by default unavailable for
OTP libs, due to the fact that shipped .beam files contain source
paths valid on the build host, not on the local machine.