Make it clear that the current format is not elixir_docs_v1
The original idea was to ultimately use Elixir's format,
but it has not happened until now - let's make it clear.
Now, once EEP-48 is out and a new format is documented,
it's better to implement the new one which Elixir should also
eventually support, while leaving the old docsh one as is.
Moreover, don't use the ExDc chunk anymore - switch to Docs instead.